+++ examples/utils/run_TransferBoolVar.py
@@ -1,27 +1,33 @@
 #!/usr/bin/env python
-# Created by "Thieu" at 13:34, 27/09/2023 ----------%                                                                               
-#       Email: nguyenthieu2102@gmail.com            %                                                    
-#       Github: https://github.com/thieu1995        %                         
+# Created by "Thieu" at 15:07, 05/11/2023 ----------%
+#       Email: nguyenthieu2102@gmail.com            %
+#       Github: https://github.com/thieu1995        %
 # --------------------------------------------------%
 
 import numpy as np
-from mealpy import IntegerVar, Problem
+from mealpy import TransferBoolVar, Problem
 
+print(f"Supported transfer functions: {TransferBoolVar.SUPPORTED_TF_FUNCS}")
 
-# ## 1. One variable
+# ## 1. Show error
 # bounds = [
-#     IntegerVar(lb=-10, ub=10, name="delta"),
+#     TransferBoolVar(n_vars=-2, name="delta", tf_func="vstf_01"),
 # ]
 
-## 2. Multiple variables
+## 2. One variable
+# bounds = [
+#     TransferBoolVar(n_vars=1, name="delta", tf_func="vstf_01"),
+# ]
+
+# ## 3. Multiple variables
 bounds = [
-    IntegerVar(lb=(-1, )*20, ub=(1, )*20, name="delta")
+    TransferBoolVar(n_vars=11, name="delta", tf_func="sstf_02"),
 ]
+
 
 problem = Problem(bounds, obj_func=lambda sol: np.sum(sol**2))
 print(f"Problem: {problem}")
 print(f"Bounds: {problem.bounds}")
-print(problem.lb, problem.ub)
 
 ## Generate encoded solution (the real-value solution)
 x = problem.generate_solution()
@@ -35,3 +41,11 @@ print(f"Real value solution: {x}")
 print(f"Encoded solution: {x1}")
 print(f"Bounded solution: {x2}")
 print(f"Real value solution after decoded: {x3}")
+
+## Supported feature selection by setting all_zeros=False
+bounds = [
+    TransferBoolVar(n_vars=2, name="delta", tf_func="sstf_02"),
+]
+problem = Problem(bounds, obj_func=lambda sol: np.sum(sol**2))
+print(problem.generate_solution(encoded=False))
+## It will never generate a solution with all variables = 0.